[BACKEND] Move mok entities to other package
