+++ app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/configuration/persistence/PersistenceMokConfig.java
@@ -14,62 +14,62 @@ import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
-import pl.lodz.p.it.inz.sgruda.multiStore.entities.AccessLevelEntity;
-import pl.lodz.p.it.inz.sgruda.multiStore.entities.AccountEntity;
-import pl.lodz.p.it.inz.sgruda.multiStore.entities.AuthenticationDataEntity;
-import pl.lodz.p.it.inz.sgruda.multiStore.entities.ForgotPasswordTokenEntity;
+import pl.lodz.p.it.inz.sgruda.multiStore.entities.mok.AccessLevelEntity;
+import pl.lodz.p.it.inz.sgruda.multiStore.entities.mok.AccountEntity;
+import pl.lodz.p.it.inz.sgruda.multiStore.entities.mok.AuthenticationDataEntity;
+import pl.lodz.p.it.inz.sgruda.multiStore.entities.mok.ForgotPasswordTokenEntity;
 
 import javax.sql.DataSource;
 
 
-@Configuration
-@EnableTransactionManagement
-@EnableJpaRepositories(basePackages = "pl.lodz.p.it.inz.sgruda.multiStore.mok.repositories",
-        entityManagerFactoryRef = "mokManagerFactory",
-        transactionManagerRef= "mokTransactionManager"
-)
+//@Configuration
+//@EnableTransactionManagement
+//@EnableJpaRepositories(basePackages = "pl.lodz.p.it.inz.sgruda.multiStore.mok.repositories",
+//        entityManagerFactoryRef = "mokManagerFactory",
+//        transactionManagerRef= "mokTransactionManager"
+//)
 public class PersistenceMokConfig {
-    @Value("${spring.datasource.mok.url}")
-    private String url;
-    @Value("${spring.datasource.mok.username}")
-    private String username;
-    @Value("${spring.datasource.mok.password}")
-    private String password;
-    @Value("${spring.datasource.mok.driverClassName}")
-    private String driver;
-
-    @Bean
-    @Primary
-//    @ConfigurationProperties("spring.datasource.mok")
-    public DataSourceProperties mokDataSourceProperties() {
-        DataSourceProperties dataSourceProperties = new DataSourceProperties();
-        dataSourceProperties.setUrl(url);
-        dataSourceProperties.setUsername(username);
-        dataSourceProperties.setPassword(password);
-        dataSourceProperties.setDriverClassName(driver);
-        return dataSourceProperties;
-    }
-
-    @Bean
-    @Primary
-    @ConfigurationProperties("spring.datasource.mok.configuration")
-    public DataSource mokDataSource() {
-        return mokDataSourceProperties().initializeDataSourceBuilder()
-                .type(HikariDataSource.class).build();
-    }
-    @Primary
-    @Bean(name = "mokManagerFactory")
-    public LocalContainerEntityManagerFactoryBean mokManagerFactory(EntityManagerFactoryBuilder builder) {
-        return builder
-                .dataSource(mokDataSource())
-                .packages(AccountEntity.class, AccessLevelEntity.class, AuthenticationDataEntity.class, ForgotPasswordTokenEntity.class)
-                .build();
-    }
-
-    @Primary
-    @Bean
-    public PlatformTransactionManager mokTransactionManager(
-            final @Qualifier("mokManagerFactory") LocalContainerEntityManagerFactoryBean mokManagerFactory) {
-        return new JpaTransactionManager(mokManagerFactory.getObject());
-    }
+//    @Value("${spring.datasource.mok.url}")
+//    private String url;
+//    @Value("${spring.datasource.mok.username}")
+//    private String username;
+//    @Value("${spring.datasource.mok.password}")
+//    private String password;
+//    @Value("${spring.datasource.mok.driverClassName}")
+//    private String driver;
+//
+//    @Bean
+//    @Primary
+////    @ConfigurationProperties("spring.datasource.mok")
+//    public DataSourceProperties mokDataSourceProperties() {
+//        DataSourceProperties dataSourceProperties = new DataSourceProperties();
+//        dataSourceProperties.setUrl(url);
+//        dataSourceProperties.setUsername(username);
+//        dataSourceProperties.setPassword(password);
+//        dataSourceProperties.setDriverClassName(driver);
+//        return dataSourceProperties;
+//    }
+//
+//    @Bean
+//    @Primary
+//    @ConfigurationProperties("spring.datasource.mok.configuration")
+//    public DataSource mokDataSource() {
+//        return mokDataSourceProperties().initializeDataSourceBuilder()
+//                .type(HikariDataSource.class).build();
+//    }
+//    @Primary
+//    @Bean(name = "mokManagerFactory")
+//    public LocalContainerEntityManagerFactoryBean mokManagerFactory(EntityManagerFactoryBuilder builder) {
+//        return builder
+//                .dataSource(mokDataSource())
+//                .packages(AccountEntity.class, AccessLevelEntity.class, AuthenticationDataEntity.class, ForgotPasswordTokenEntity.class)
+//                .build();
+//    }
+//
+//    @Primary
+//    @Bean
+//    public PlatformTransactionManager mokTransactionManager(
+//            final @Qualifier("mokManagerFactory") LocalContainerEntityManagerFactoryBean mokManagerFactory) {
+//        return new JpaTransactionManager(mokManagerFactory.getObject());
+//    }
 }
