+++ app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/dto/mappers/mok/AuthenticationDataMapper.java
@@ -2,8 +2,8 @@ package pl.lodz.p.it.inz.sgruda.multiStore.dto.mappers.mok;
 
 import pl.lodz.p.it.inz.sgruda.multiStore.dto.mappers.Mapper;
 import pl.lodz.p.it.inz.sgruda.multiStore.dto.mok.AuthenticationDataDTO;
-import pl.lodz.p.it.inz.sgruda.multiStore.entities.AccountEntity;
-import pl.lodz.p.it.inz.sgruda.multiStore.entities.AuthenticationDataEntity;
+import pl.lodz.p.it.inz.sgruda.multiStore.entities.mok.AccountEntity;
+import pl.lodz.p.it.inz.sgruda.multiStore.entities.mok.AuthenticationDataEntity;
 import pl.lodz.p.it.inz.sgruda.multiStore.utils.HashMethod;
