+++ app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/mok/repositories/AccessLevelRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import pl.lodz.p.it.inz.sgruda.multiStore.entities.AccessLevelEntity;
+import pl.lodz.p.it.inz.sgruda.multiStore.entities.mok.AccessLevelEntity;
 import pl.lodz.p.it.inz.sgruda.multiStore.utils.enums.RoleName;
 
 import java.util.Optional;
