+++ app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/mok/services/interfaces/AuthService.java
@@ -1,6 +1,6 @@
 package pl.lodz.p.it.inz.sgruda.multiStore.mok.services.interfaces;
 
-import pl.lodz.p.it.inz.sgruda.multiStore.entities.AccountEntity;
+import pl.lodz.p.it.inz.sgruda.multiStore.entities.mok.AccountEntity;
 import pl.lodz.p.it.inz.sgruda.multiStore.exceptions.mok.EmailAlreadyExistsException;
 import pl.lodz.p.it.inz.sgruda.multiStore.exceptions.mok.UsernameAlreadyExistsException;
