+++ app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/mok/services/interfaces/OAuth2Service.java
@@ -1,7 +1,7 @@
 package pl.lodz.p.it.inz.sgruda.multiStore.mok.services.interfaces;
 
 import org.springframework.security.oauth2.core.user.OAuth2User;
-import pl.lodz.p.it.inz.sgruda.multiStore.entities.AccountEntity;
+import pl.lodz.p.it.inz.sgruda.multiStore.entities.mok.AccountEntity;
 import pl.lodz.p.it.inz.sgruda.multiStore.exceptions.mok.AccountNotExistsException;
 
 public interface OAuth2Service {
