+++ app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/security/oauth2/CustomOAuth2UserService.java
@@ -9,7 +9,7 @@ import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
-import pl.lodz.p.it.inz.sgruda.multiStore.entities.AccountEntity;
+import pl.lodz.p.it.inz.sgruda.multiStore.entities.mok.AccountEntity;
 import pl.lodz.p.it.inz.sgruda.multiStore.exceptions.mok.AccountNotExistsException;
 import pl.lodz.p.it.inz.sgruda.multiStore.mok.services.interfaces.OAuth2Service;
 import pl.lodz.p.it.inz.sgruda.multiStore.security.UserPrincipal;
